requirements.txt
fastapi
uvicorn
python-multipart
openai
faiss-cpu
numpy
python-dotenv
